reference-app/backend/requirements.txt
Flask
requests
pylint
gunicorn
pymongo
Flask-PyMongo
prometheus-client
opentelemetry-instrumentation
opentelemetry-sdk
prometheus-flask-exporter
jaeger-client
opentelemetry-api
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-request
Flask-OpenTracing